despawn field
Created a despawn field behind the camera which should in theory delete objects that pass the camera frame.

no idea why this doesnt work. Originally i had it under the camera object so it would move with the camera but it wasnt deleting anything.

After moving it out it still didnt delete anything